Split out the rest of the commands
